Implemented thread-safe packet queuing
Disabled the timer interval from calling GetNextPacket() and instead used Pcap's OnPacketArrival. In addition, a packet processing thread will run in the background and handle buffering the packets in its own time. All packets that arrive are pushed to the packet queue and are then processed in the background thread, both of which are thread-safe now.